unk8sify a pps server test
