+++ src/server/pps/server/api_server_integration_test.go
@@ -11,10 +11,8 @@ import (
 	"google.golang.org/protobuf/encoding/protojson"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/pachyderm/pachyderm/v2/src/auth"
 	"github.com/pachyderm/pachyderm/v2/src/pfs"
 	"github.com/pachyderm/pachyderm/v2/src/pps"
-	at "github.com/pachyderm/pachyderm/v2/src/server/auth/server/testing"
 
 	"github.com/pachyderm/pachyderm/v2/src/internal/client"
 	"github.com/pachyderm/pachyderm/v2/src/internal/errors"
@@ -268,55 +266,4 @@ func TestSidecarMetrics(t *testing.T) {
 		require.True(t, foundMetrics, "metrics port must exist")
 	}
 	require.True(t, foundSidecar, "sidecar must exist")
-}
-
-func TestProjectDefaultsCreatedBy(t *testing.T) {
-	t.Parallel()
-	c := at.EnvWithAuth(t).PachClient
-
-	admin := tu.AuthenticateClient(t, c, auth.RootUser)
-	aliceName, alice := tu.RandomRobot(t, c, "alice")
-	adminProject := tu.UniqueString("adminProject")
-	require.NoError(t, admin.CreateProject(adminProject))
-	_, err := admin.PpsAPIClient.SetProjectDefaults(admin.Ctx(), &pps.SetProjectDefaultsRequest{
-		Project:             &pfs.Project{Name: adminProject},
-		ProjectDefaultsJson: `{"createPipelineRequest": {"autoscaling": true}}`,
-	})
-	require.NoError(t, err)
-	resp, err := admin.PpsAPIClient.GetProjectDefaults(c.Ctx(), &pps.GetProjectDefaultsRequest{Project: &pfs.Project{Name: adminProject}})
-	require.NoError(t, err)
-	require.Equal(t, "pach:root", resp.CreatedBy)
-
-	aliceProject := tu.UniqueString("aliceProject")
-	require.NoError(t, alice.CreateProject(aliceProject))
-	_, err = alice.PpsAPIClient.SetProjectDefaults(alice.Ctx(), &pps.SetProjectDefaultsRequest{
-		Project:             &pfs.Project{Name: aliceProject},
-		ProjectDefaultsJson: `{"createPipelineRequest": {"autoscaling": true}}`,
-	})
-	require.NoError(t, err)
-	resp, err = alice.PpsAPIClient.GetProjectDefaults(c.Ctx(), &pps.GetProjectDefaultsRequest{Project: &pfs.Project{Name: aliceProject}})
-	require.NoError(t, err)
-	require.Equal(t, aliceName, resp.CreatedBy)
-
-	pi, err := admin.InspectProject(adminProject)
-	require.NoError(t, err)
-	require.Equal(t, "pach:root", pi.CreatedBy)
-	pi, err = alice.InspectProject(aliceProject)
-	require.NoError(t, err)
-	require.Equal(t, aliceName, pi.CreatedBy)
-
-	pp, err := admin.ListProject()
-	require.NoError(t, err)
-	for _, p := range pp {
-		switch p.Project.Name {
-		case "default":
-			require.Equal(t, "", p.CreatedBy)
-		case adminProject:
-			require.Equal(t, "pach:root", p.CreatedBy)
-		case aliceProject:
-			require.Equal(t, aliceName, p.CreatedBy)
-		default:
-			t.Fatalf("unexpected project %v", p.Project)
-		}
-	}
 }
